terraform/module/cloud_run_job/requirements.txt
fastapi
uvicorn
google-cloud-pubsub
faker
unidecode

